ERC20 tokens data is showed at wallet and walletScreen pages

+++ currencies/erc20TokenData.go
@@ -6,5 +6,5 @@ import (
 type Erc20TokenData struct {
 	Name string
 	Symbol string
-	Decimals int64
+	Decimals int
 }
